@@ -2,19 +2,22 @@ import 'dart:math';
 
 import 'package:toplife/core/utils/chance.dart';
 import 'package:toplife/core/utils/numbers/get_negative_or_positive_multiplier.dart';
-import 'package:toplife/main_systems/system_shop_and_storage/constants/car_problem.dart';
+import 'package:toplife/main_systems/system_shop_and_storage/constants/car_problem_type.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/constants/car_quality.dart';
 import 'package:toplife/core/data_source/drift_database/database_provider.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/domain/repository/car_repository.dart';
+import 'package:toplife/main_systems/system_shop_and_storage/domain/usecases/car/car_problem/create_car_problem_usecase.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/domain/usecases/car/get_car_max_condition.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/util/get_car_quality_enum.dart';
 
 class DriveCarUsecase {
   final CarRepository _carRepository;
+  final CreateCarProblemUsecase _createCarProblemUsecase;
   final GetCarMaxConditionUsecase _getCarMaxConditionUsecase;
 
   const DriveCarUsecase(
     this._carRepository,
+    this._createCarProblemUsecase,
     this._getCarMaxConditionUsecase,
   );
 
@@ -47,32 +50,37 @@ class DriveCarUsecase {
     late final int chanceOfProblem;
 
     if (maxCondition <= 20) {
-      chanceOfProblem = 50;
+      chanceOfProblem = 25;
     } else if (maxCondition <= 50) {
-      chanceOfProblem = 20;
-    } else {
       chanceOfProblem = 10;
+    } else {
+      chanceOfProblem = 5;
     }
 
     final bool carGetsProblem = Chance.getTrueOrFalseBasedOnPercentageChance(
       trueChancePercentage: chanceOfProblem,
     );
 
-    late final CarProblem newCarProblem;
+    late final CarProblemType newCarProblemType;
     if (carGetsProblem) {
-      newCarProblem = CarProblem.values[Random().nextInt(
-        CarProblem.values.length,
+      //choose random problem
+      newCarProblemType = CarProblemType.values[Random().nextInt(
+        CarProblemType.values.length,
       )];
-    } else {
-      newCarProblem = CarProblem.noProblem;
+
+      //add chosen problem
+      await _createCarProblemUsecase.execute(
+        car: car,
+        currentDay: currentDay,
+        carProblemType: newCarProblemType,
+      );
     }
 
-    //update car
+    // update car
     await _carRepository.updateCar(
       car.copyWith(
         useCondition: newUseCondition,
         fuelTank: newFuelTank,
-        problem: newCarProblem.name,
       ),
     );
   }
