@@ -1,3 +1,4 @@
+import 'package:toplife/main_systems/system_shop_and_storage/data/repository/car_problem_repository_impl.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/data/repository/car_repository_impl.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/data/repository/current_home_repository_impl.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/data/repository/food_repository_impl.dart';
@@ -6,6 +7,7 @@ import 'package:toplife/main_systems/system_shop_and_storage/data/repository/hou
 import 'package:toplife/main_systems/system_shop_and_storage/data/repository/item_repository_impl.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/data/repository/jewelry_repository_impl.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/domain/dao/car_dao.dart';
+import 'package:toplife/main_systems/system_shop_and_storage/domain/dao/car_problem_dao.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/domain/dao/current_home_dao.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/domain/dao/food_dao.dart';
 import 'package:toplife/main_systems/system_shop_and_storage/domain/dao/fridge_food_dao.dart';
@@ -24,6 +26,7 @@ class ShopAndStorageRepositories {
   final StoreroomItemDao _storeroomItemDao;
   final JewelryDao _jewelryDao;
   final CurrentHomeDao _currentHomeDao;
+  final CarProblemDao _carProblemDao;
 
   const ShopAndStorageRepositories({
     required CarDao carDao,
@@ -34,6 +37,7 @@ class ShopAndStorageRepositories {
     required StoreroomItemDao storeroomItemDao,
     required JewelryDao jewelryDao,
     required CurrentHomeDao currentHomeDao,
+    required CarProblemDao carProblemDao,
   })  : _carDao = carDao,
         _houseDao = houseDao,
         _foodDao = foodDao,
@@ -41,7 +45,8 @@ class ShopAndStorageRepositories {
         _itemDao = itemDao,
         _storeroomItemDao = storeroomItemDao,
         _jewelryDao = jewelryDao,
-        _currentHomeDao = currentHomeDao;
+        _currentHomeDao = currentHomeDao,
+        _carProblemDao = carProblemDao;
 
   CarRepositoryImpl get carRepositoryImpl => CarRepositoryImpl(carDao: _carDao);
 
@@ -65,4 +70,7 @@ class ShopAndStorageRepositories {
 
   CurrentHomeRepositoryImpl get currentHomeRepositoryImpl =>
       CurrentHomeRepositoryImpl(currentHomeDao: _currentHomeDao);
+
+  CarProblemRepositoryImpl get carProblemRepositoryImpl =>
+      CarProblemRepositoryImpl(carProblemDao: _carProblemDao);
 }
