@@ -33,10 +33,11 @@ void main() {
       currentState: "Ontario",
       currentCountry: "Canada",
       money: 120,
-        emotionalState: "normal",
+      emotionalState: "normal",
       zodiacSign: "",
       hasDriversLicense: true,
       transportMode: "bus",
+      drivingMode: "drivingMode",
       hasFertilityIssues: false,
       onBirthControl: false,
       isSterile: false,
