@@ -1,71 +1,67 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:toplife/main_systems/system_shop_and_storage/constants/car_problem.dart';
+import 'package:toplife/main_systems/system_shop_and_storage/constants/car_problem_type.dart';
 
 void main() {
-  group("Car problem:", () {
-    test("no problem has the correct fullname", () {
-      expect(CarProblem.noProblem.fullName, "");
-    });
-
+  group("Car problem type:", () {
     test("sputtering engine has the correct fullname", () {
       expect(
-        CarProblem.sputteringEngine.fullName,
+        CarProblemType.sputteringEngine.fullName,
         "Sputtering Engine",
       );
     });
 
     test("alternator failure has the correct fullname", () {
       expect(
-        CarProblem.alternatorFailure.fullName,
+        CarProblemType.alternatorFailure.fullName,
         "Alternator Failure",
       );
     });
 
     test("faulty fuse has the correct fullname", () {
       expect(
-        CarProblem.faultyFuse.fullName,
+        CarProblemType.faultyFuse.fullName,
         "Faulty Fuse",
       );
     });
 
     test("damaged oxygen sensors has the correct fullname", () {
       expect(
-        CarProblem.damagedOxygenSensors.fullName,
+        CarProblemType.damagedOxygenSensors.fullName,
         "Damaged Oxygen Sensors",
       );
     });
 
     test("dead battery has the correct fullname", () {
       expect(
-        CarProblem.deadBattery.fullName,
+        CarProblemType.deadBattery.fullName,
         "Dead Battery",
       );
     });
 
     test("faulty brakes has the correct fullname", () {
       expect(
-        CarProblem.faultyBrakes.fullName,
+        CarProblemType.faultyBrakes.fullName,
         "Faulty Brakes",
       );
     });
 
     test("faulty spark plugs has the correct fullname", () {
       expect(
-        CarProblem.faultySparkPlugs.fullName,
+        CarProblemType.faultySparkPlugs.fullName,
         "Faulty Spark Plugs",
       );
     });
 
     test("overheating has the correct fullname", () {
       expect(
-        CarProblem.overheating.fullName,
+        CarProblemType.overheating.fullName,
         "Overheating",
       );
     });
 
     test("faulty oil pump has the correct fullname", () {
       expect(
-        CarProblem.faultyOilPump.fullName,
+        CarProblemType.faultyOilPump.fullName,
         "Faulty Oil Pump",
       );
     });
@@ -75,7 +71,7 @@ void main() {
     //*The moment prices are added this should fail so you need to:
     //! change it to check that all base prices are greater than the least priced item.
     test("All car problems have a base price of 0", () {
-      for (var problem in CarProblem.values) {
+      for (var problem in CarProblemType.values) {
         expect(
           problem.basePrice,
           0,
